+++ src/main/java/com/example/green/domain/example/tools/DbInitializer.java
@@ -27,9 +27,9 @@ import com.example.green.domain.point.entity.vo.PointAmount;
 import com.example.green.domain.point.entity.vo.PointSource;
 import com.example.green.domain.point.entity.vo.TargetType;
 import com.example.green.domain.point.repository.PointTransactionRepository;
-import com.example.green.domain.pointshop.entity.delivery.DeliveryAddress;
-import com.example.green.domain.pointshop.entity.delivery.vo.Address;
-import com.example.green.domain.pointshop.entity.delivery.vo.Recipient;
+import com.example.green.domain.pointshop.delivery.entity.DeliveryAddress;
+import com.example.green.domain.pointshop.delivery.entity.vo.Address;
+import com.example.green.domain.pointshop.delivery.entity.vo.Recipient;
 import com.example.green.domain.pointshop.order.entity.Order;
 import com.example.green.domain.pointshop.order.entity.OrderItem;
 import com.example.green.domain.pointshop.order.entity.vo.DeliveryAddressSnapshot;
@@ -41,7 +41,7 @@ import com.example.green.domain.pointshop.entity.pointproduct.vo.Code;
 import com.example.green.domain.pointshop.entity.pointproduct.vo.Media;
 import com.example.green.domain.pointshop.entity.pointproduct.vo.Price;
 import com.example.green.domain.pointshop.entity.pointproduct.vo.Stock;
-import com.example.green.domain.pointshop.repository.DeliveryAddressRepository;
+import com.example.green.domain.pointshop.delivery.repository.DeliveryAddressRepository;
 import com.example.green.domain.pointshop.order.repository.OrderRepository;
 import com.example.green.domain.pointshop.repository.PointProductRepository;
