--- src/main/java/com/example/green/domain/pointshop/client/PhoneVerificationClient.java
@@ -1,10 +1,0 @@
-package com.example.green.domain.pointshop.client;
-
-public interface PhoneVerificationClient {
-
-	/*
-		MVP 에서 우선, 전화 인증이 필요 없어진 상태
-	 */
-	@Deprecated
-	boolean isAuthenticated(String phoneNumber);
-}
