--- src/main/java/com/example/green/domain/pointshop/exception/deliveryaddress/DeliveryAddressException.java
@@ -1,10 +1,0 @@
-package com.example.green.domain.pointshop.exception.deliveryaddress;
-
-import com.example.green.global.error.exception.BusinessException;
-import com.example.green.global.error.exception.ExceptionMessage;
-
-public class DeliveryAddressException extends BusinessException {
-	public DeliveryAddressException(ExceptionMessage exceptionMessage) {
-		super(exceptionMessage);
-	}
-}
