+++ src/main/java/com/example/green/domain/pointshop/order/service/OrderService.java
@@ -5,8 +5,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.green.domain.pointshop.exception.OrderException;
-import com.example.green.domain.pointshop.exception.OrderExceptionMessage;
 import com.example.green.domain.pointshop.order.client.PointSpendClient;
 import com.example.green.domain.pointshop.order.client.dto.PointSpendRequest;
 import com.example.green.domain.pointshop.order.entity.Order;
@@ -14,9 +12,11 @@ import com.example.green.domain.pointshop.order.entity.OrderItem;
 import com.example.green.domain.pointshop.order.entity.vo.DeliveryAddressSnapshot;
 import com.example.green.domain.pointshop.order.entity.vo.ItemSnapshot;
 import com.example.green.domain.pointshop.order.entity.vo.MemberSnapshot;
+import com.example.green.domain.pointshop.order.exception.OrderException;
+import com.example.green.domain.pointshop.order.exception.OrderExceptionMessage;
 import com.example.green.domain.pointshop.order.repository.OrderRepository;
 import com.example.green.domain.pointshop.order.service.command.SingleOrderCommand;
-import com.example.green.domain.pointshop.service.DeliveryAddressService;
+import com.example.green.domain.pointshop.delivery.service.DeliveryAddressService;
 import com.example.green.domain.pointshop.service.PointProductService;
 
 import lombok.RequiredArgsConstructor;
