--- src/main/java/com/example/green/domain/pointshop/service/command/DeliveryAddressCreateCommand.java
@@ -1,16 +1,0 @@
-package com.example.green.domain.pointshop.service.command;
-
-import com.example.green.domain.pointshop.entity.delivery.DeliveryAddress;
-import com.example.green.domain.pointshop.entity.delivery.vo.Address;
-import com.example.green.domain.pointshop.entity.delivery.vo.Recipient;
-
-public record DeliveryAddressCreateCommand(
-	Long recipientId,
-	Recipient recipient,
-	Address address
-) {
-
-	public DeliveryAddress toDeliveryAddress() {
-		return DeliveryAddress.create(recipientId, recipient, address);
-	}
-}
