+++ src/test/java/com/example/green/domain/pointshop/order/entity/OrderTest.java
@@ -9,11 +9,11 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import com.example.green.domain.pointshop.exception.OrderException;
-import com.example.green.domain.pointshop.exception.OrderExceptionMessage;
 import com.example.green.domain.pointshop.order.entity.vo.DeliveryAddressSnapshot;
 import com.example.green.domain.pointshop.order.entity.vo.MemberSnapshot;
 import com.example.green.domain.pointshop.order.entity.vo.OrderDeliveryStatus;
+import com.example.green.domain.pointshop.order.exception.OrderException;
+import com.example.green.domain.pointshop.order.exception.OrderExceptionMessage;
 
 class OrderTest {
